+++ tests/test_gaussian.py
@@ -93,7 +93,7 @@ class TestGaussian:
         dev = qml.device("strawberryfields.gaussian", wires=2, cutoff_dim=15)
         assert dev.num_wires == 2
         assert dev.hbar == 2
-        assert dev.shots == 1000
+        assert dev.shots is None
         assert dev.cutoff == 15
         assert dev.short_name == "strawberryfields.gaussian"
 
@@ -601,8 +601,8 @@ class TestVariance:
         assert np.allclose(var, expected, atol=tol, rtol=0)
 
         # circuit jacobians
-        gradA = circuit.jacobian([r, phi], method="A")
-        gradF = circuit.jacobian([r, phi], method="F")
+        gradA = circuit.qtape.jacobian(dev, method="analytic")
+        gradF = circuit.qtape.jacobian(dev, method="numeric")
         expected = np.array(
             [
                 2 * np.exp(2 * r) * np.sin(phi) ** 2 - 2 * np.exp(-2 * r) * np.cos(phi) ** 2,
@@ -630,7 +630,7 @@ class TestVariance:
         assert np.allclose(var, expected, atol=tol, rtol=0)
 
         # circuit jacobians
-        gradF = circuit.jacobian([n, a], method="F")
+        gradF = circuit.qtape.jacobian(dev, method="numeric")
         expected = np.array([2 * a ** 2 + 2 * n + 1, 2 * a * (2 * n + 1)])
         assert np.allclose(gradF, expected, atol=tol, rtol=0)
 
@@ -717,13 +717,21 @@ class TestProbability:
 
         n = np.arange(cutoff)
 
+        # construct tape
+        circuit.construct([a, phi], {})
+
         # differentiate with respect to parameter a
-        res_F = circuit.jacobian([a, phi], wrt={0}, method="F").flat
+        circuit.qtape.trainable_params = {0}
+        res_F = circuit.qtape.jacobian(dev, method="numeric").flatten()
         expected_gradient = 2 * np.exp(-(a ** 2)) * a ** (2 * n - 1) * (n - a ** 2) / fac(n)
         assert np.allclose(res_F, expected_gradient, atol=tol, rtol=0)
 
+        # re-construct tape to reset trainable_params
+        circuit.construct([a, phi], {})
+
         # differentiate with respect to parameter phi
-        res_F = circuit.jacobian([a, phi], wrt={1}, method="F").flat
+        circuit.qtape.trainable_params = {1}
+        res_F = circuit.qtape.jacobian(dev, method="numeric").flatten()
         expected_gradient = 0
         assert np.allclose(res_F, expected_gradient, atol=tol, rtol=0)
 
@@ -731,6 +739,7 @@ class TestProbability:
         """Test that the jacobian of the probability for a squeezed states is
         approximated well with finite differences"""
         cutoff = 5
+        # import pdb; pdb.set_trace()
 
         dev = qml.device("strawberryfields.gaussian", wires=1, cutoff_dim=cutoff)
 
@@ -744,8 +753,12 @@ class TestProbability:
 
         n = np.arange(cutoff)
 
+        # construct tape
+        circuit.construct([r, phi], {})
+
         # differentiate with respect to parameter r
-        res_F = circuit.jacobian([r, phi], wrt={0}, method="F").flatten()
+        circuit.qtape.trainable_params = {0}
+        res_F = circuit.qtape.jacobian(dev, method="numeric").flatten()
         assert res_F.shape == (cutoff,)
 
         expected_gradient = (
@@ -757,26 +770,30 @@ class TestProbability:
         expected_gradient[n % 2 != 0] = 0
         assert np.allclose(res_F, expected_gradient, atol=tol, rtol=0)
 
+        # re-construct tape to reset trainable_params
+        circuit.construct([r, phi], {})
+
         # differentiate with respect to parameter phi
-        res_F = circuit.jacobian([r, phi], wrt={1}, method="F").flat
+        circuit.qtape.trainable_params = {1}
+        res_F = circuit.qtape.jacobian(dev, method="numeric").flatten()
         expected_gradient = 0
         assert np.allclose(res_F, expected_gradient, atol=tol, rtol=0)
 
     def test_finite_diff_coherent_two_wires(self, tol):
-        """Test that the jacobian of the probability for a coherent states is
-        approximated well with finite differences"""
+        """Test that the jacobian of the probability for a coherent states is approximated well with
+        finite differences"""
         cutoff = 4
 
-        dev = qml.device("strawberryfields.fock", wires=2, cutoff_dim=cutoff)
+        dev = qml.device("strawberryfields.gaussian", wires=2, cutoff_dim=cutoff)
 
-        @qml.qnode(dev)
+        @qml.qnode(dev, diff_method="finite-diff")
         def circuit(a, phi):
             qml.Displacement(a, phi, wires=0)
             qml.Displacement(a, phi, wires=1)
             return qml.probs(wires=[0, 1])
 
-        a = 0.4
-        phi = -0.12
+        a = np.array(0.4, requires_grad=True)
+        phi = np.array(-0.12, requires_grad=False)
 
         c = np.arange(cutoff)
         d = np.arange(cutoff)
@@ -785,7 +802,7 @@ class TestProbability:
         n1 = n1.flatten()
 
         # differentiate with respect to parameter a
-        res_F = circuit.jacobian([a, phi], wrt={0}, method="F").flat
+        res_F = qml.jacobian(circuit)(a, phi)
         expected_gradient = (
             2
             * (a ** (-1 + 2 * n0 + 2 * n1))
@@ -796,7 +813,10 @@ class TestProbability:
         assert np.allclose(res_F, expected_gradient, atol=tol, rtol=0)
 
         # differentiate with respect to parameter phi
-        res_F = circuit.jacobian([a, phi], wrt={1}, method="F").flat
+        a = np.array(0.4, requires_grad=False)
+        phi = np.array(-0.12, requires_grad=True)
+
+        res_F = qml.jacobian(circuit)(a, phi)
         expected_gradient = 0
         assert np.allclose(res_F, expected_gradient, atol=tol, rtol=0)
 
@@ -813,8 +833,11 @@ class TestProbability:
         a = 0.4
         phi = -0.12
 
+        # construct tape
+        circuit.construct([a, phi], {})
+
         with pytest.raises(ValueError, match="The analytic gradient method cannot be used with"):
-            res_F = circuit.jacobian([a, phi], wrt={0}, method="A").flat
+            _ = circuit.qtape.jacobian(dev, method="analytic").flatten()
 
     def test_tensorn_one_mode_is_mean_photon(self, tol):
         """Test variance of TensorN for a single mode, which resorts to
@@ -841,6 +864,6 @@ class TestProbability:
         assert np.allclose(var, expected, atol=tol, rtol=0)
 
         # circuit jacobians
-        gradF = circuit.jacobian([n, a], method="F")
+        gradF = circuit.qtape.jacobian(dev, method="numeric")
         expected = np.array([2 * a ** 2 + 2 * n + 1, 2 * a * (2 * n + 1)])
         assert np.allclose(gradF, expected, atol=tol, rtol=0)